fix# 上海外服健康点展示问题

+++ ehealth-recipe-api/src/main/java/com/ngari/recipe/recipeorder/model/RecipeOrderBean.java
@@ -342,7 +342,7 @@ public class RecipeOrderBean implements Serializable {
     private Integer thirdPayType;
 
     @ItemProperty(alias = "第三方支付金额")
-    private BigDecimal thirdPayFee;
+    private double thirdPayFee;
     @ItemProperty(alias = "预约取药开始时间")
     private String expectStartTakeTime;
     @ItemProperty(alias = "预约取药结束时间")
@@ -471,11 +471,11 @@ public class RecipeOrderBean implements Serializable {
         this.thirdPayType = thirdPayType;
     }
 
-    public BigDecimal getThirdPayFee() {
+    public double getThirdPayFee() {
         return thirdPayFee;
     }
 
-    public void setThirdPayFee(BigDecimal thirdPayFee) {
+    public void setThirdPayFee(double thirdPayFee) {
         this.thirdPayFee = thirdPayFee;
     }
